added syntax hide
"hide" isn't a command but is a syntax to remove the current function from the auto complete. When typing /function it would hide that function from the list if the string hide was present in that function. I keep getting errors for this when I use it in vs code, so I thought I would make a commit.